requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.5
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
emoji==2.11.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.56
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.8.1
numpy==1.26.4
openai==1.27.0
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pickleshare==0.7.5
pillow==10.3.0
pinecone-client==3.2.2
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
streamlit==1.34.0
streamlit-chat==0.1.1
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.6.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.13.7
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
